fix for PlacementsDetail to be consistent "at least 1"
(cherry picked from commit bee5ffd2ab642535cc873067be744e246f9231a6)
(cherry picked from commit 079a8ad259422ebaa8df1ff79455ef3ff76af7cb)

+++ src/Sfa.Tl.Matching.Models/ViewModel/SearchResultsViewModelItem.cs
@@ -7,7 +7,7 @@ namespace Sfa.Tl.Matching.Models.ViewModel
         public int ProviderVenueId { get; set; }
         public bool IsSelected { get; set; }
         public string Postcode { get; set; }
-        public decimal? Distance { get; set; }
+        public double? Distance { get; set; }
         public string ProviderName { get; set; }
         public IEnumerable<string> QualificationShortTitles { get; set; }
     }
